Change id types from long to int

+++ src/main/java/com/example/meven/arpentage_api/repository/MemberRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface MemberRepository extends CrudRepository<Member, Long> {
+public interface MemberRepository extends CrudRepository<Member, Integer> {
 }
